requirements.txt
tensorflow-cpu
keras